requirements.txt
gradio==5.25.2
pandas==2.2.2
openpyxl==3.1.5
pillow==11.1.0
pyngrok==7.2.4
